requirements.txt
Django==5.2
djangorestframework==3.16.0
rich==14.0.0